+++ src/main/java/com/mentalfrostbyte/jello/module/impl/gui/jello/Compass.java
@@ -28,7 +28,7 @@ public class Compass extends Module {
                 int var5 = 60;
                 int var6 = !Minecraft.getInstance().gameSettings.showDebugInfo ? 0 : 60;
                 List<Integer> var7 = this.method16660((int) this.method16662(mc.player.rotationYaw), var4);
-                int var8 = (Integer) var7.get(var4);
+                int var8 = var7.get(var4);
                 if (var8 == 0 && this.method16662(mc.player.rotationYaw) > 345.0F) {
                     var8 = 360;
                 }
